Create test_optimizedqueries.py
To verify the performance of the optimized queries, I created a script that connects to the SQLite database and runs EXPLAIN QUERY PLAN on the new version of the monthly sales query that uses the sale_month column. The script executes the query SELECT SUM(quantity * sale_price) FROM sales WHERE sale_month = '2025-04'; and prints out the execution plan results. This allows me to check whether the database is using the index on sale_month and helps confirm that the rewritten query is more efficient.